Move preassembly manager to preassemble_db.
